Refactored type names and extracted them to files.
