+++ src/common/services/sirv-cdn/sirv-cdn.types.ts
@@ -1,0 +1,16 @@
+interface SirvCdnTokenResponse {
+  token: string;
+  expiresIn: number;
+}
+
+interface SirvCdnFileData {
+  filename: string;
+  meta: FileMetaData;
+}
+
+interface FileMetaData {
+  width: number;
+  height: number;
+}
+
+export { SirvCdnFileData, SirvCdnTokenResponse };
