+++ src/common/services/slack-bot-api/slack-bot-api.types.ts
@@ -1,0 +1,12 @@
+interface SlackBotApiTokenResponse {
+  token: string;
+}
+
+interface SlackBotApiImageData {
+  id: number;
+  fileName: string;
+  x: number;
+  y: number;
+}
+
+export { SlackBotApiTokenResponse, SlackBotApiImageData };
