--- src/common/types/SirvCdnTokenResponse.ts
@@ -1,4 +1,0 @@
-export default interface SirvCdnTokenResponse {
-  token: string;
-  expiresIn: number;
-}
